requirements.txt
pyyaml==5.4.1
pyserial==3.5